+++ lib/typecheck/typecheck_unify.ml
@@ -897,7 +897,8 @@ let bind_linear_field_sentinels varname ty env =
         match repr fty with
         | TLin (lin, _) when lin <> Ast.Unrestricted ->
           let key = varname ^ "#" ^ fname in
-          let le = { le_name = key; le_lin = lin; le_used = ref false; le_first_use = ref None } in
+          let le = { le_name = key; le_lin = lin; le_used = ref false; le_first_use = ref None;
+                     le_pending = None; le_dup = ref None } in
           { acc_env with lin = le :: acc_env.lin }
         | _ -> acc_env
       ) env flds